feat: complete day05
I had a really neat solution for the sorting of the things. But the
input seemed to be somehow cursed that it didn't work. Rather than
sorting each list based on all the rules, I wanted to create a single
sorted list, and then a (filter positive? (map pages-contains sorted))
would automatically sort the thing, in what I assume must be an elegant
way. But it seemed like it was not possible to sort a list of all the
numbers present in the rules (using a bubble-sort) type of mechanic, I
assume because the rules were constructed with some kind of hole
preventing that. Which upsets me. =/